requirements.txt
mysqlclient==2.0.3
aniso8601==9.0.1
appnope==0.1.0
argon2-cffi==20.1.0
arrow==1.0.3
async-generator==1.10
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
Brotli==1.0.9
Cerberus==1.3.3
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Compress==1.9.0
Flask-Cors==3.0.10
flask-ldap3-login==0.9.16
Flask-Login==0.5.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jira==2.0.0
jsonschema==3.2.0
ldap3==2.9
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.5
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
numpy==1.20.1
oauthlib==3.1.0
packaging==20.4
palettable==3.3.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.14.3
powerline-status==2.7
prometheus-client==0.8.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.7.1
PyJWT==2.0.1
pymongo==3.11.3
pyodps==0.10.7
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.4.7
terminado==0.9.1
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
WTForms==2.3.3
parade==0.2.0.5
redis
parade-server==0.0.2.1